Add consolidation of onCancel and onConfirm methods in SupperlierRequisitionsPage

+++ src/pages/SupplierRequisitionsPage.js
@@ -45,30 +45,25 @@ export class SupplierRequisitionsPage extends React.Component {
     this.setState({ usesPrograms: !!getAllPrograms(settings, database) });
   }
 
-  onCancelByProgram = () => {
-    this.setState({ byProgramModalOpen: false });
-  };
-
-  onDeleteConfirm = () => {
+  onDelete = shouldConfirm => () => {
     const { selection } = this.state;
     const { database } = this.props;
-    database.write(() => {
-      const requisitionsToDelete = [];
-      for (let i = 0; i < selection.length; i += 1) {
-        const requisition = this.requisitions.find(
-          currentRequisition => currentRequisition.id === selection[i]
-        );
-        if (requisition.isValid() && !requisition.isFinalised) {
-          requisitionsToDelete.push(requisition);
-        }
-      }
-      database.delete('Requisition', requisitionsToDelete);
-    });
-    this.setState({ selection: [] });
-    this.refreshData();
-  };
 
-  onDeleteCancel = () => {
+    if (shouldConfirm) {
+      database.write(() => {
+        const requisitionsToDelete = [];
+        for (let i = 0; i < selection.length; i += 1) {
+          const requisition = this.requisitions.find(
+            currentRequisition => currentRequisition.id === selection[i]
+          );
+          if (requisition.isValid() && !requisition.isFinalised) {
+            requisitionsToDelete.push(requisition);
+          }
+        }
+        database.delete('Requisition', requisitionsToDelete);
+      });
+    }
+
     this.setState({ selection: [] });
     this.refreshData();
   };
@@ -77,7 +72,7 @@ export class SupplierRequisitionsPage extends React.Component {
     const { database, currentUser } = this.props;
 
     if (!requisitionValues) {
-      this.setState({ byProgramModalOpen: false });
+      this.setState({ byProgramModalOpen: false, isCreatingRequisition: false });
       return;
     }
 
@@ -94,7 +89,6 @@ export class SupplierRequisitionsPage extends React.Component {
       if (requisition.program) requisition.addItemsFromProgram(database);
     });
 
-    this.setState({ byProgramModalOpen: false });
     this.navigateToRequisition(requisition);
   };
 
@@ -104,7 +98,8 @@ export class SupplierRequisitionsPage extends React.Component {
 
   navigateToRequisition = requisition => {
     const { navigateTo } = this.props;
-    this.setState({ selection: [] }); // Clear any requsitions selected for deletion.
+    // Clear any requsitions selected to delete.
+    this.setState({ selection: [], byProgramModalOpen: false });
     navigateTo('supplierRequisition', `${navStrings.requisition} ${requisition.serialNumber}`, {
       requisition,
     });
@@ -159,14 +154,9 @@ export class SupplierRequisitionsPage extends React.Component {
 
   renderNewRequisitionButton = () => {
     const { usesPrograms } = this.state;
-    const newStateObject = usesPrograms
-      ? { byProgramModalOpen: true }
-      : { isCreatingRequisition: true };
+    const newState = usesPrograms ? { byProgramModalOpen: true } : { isCreatingRequisition: true };
     return (
-      <PageButton
-        text={buttonStrings.new_requisition}
-        onPress={() => this.setState(newStateObject)}
-      />
+      <PageButton text={buttonStrings.new_requisition} onPress={() => this.setState(newState)} />
     );
   };
 
@@ -231,8 +221,8 @@ export class SupplierRequisitionsPage extends React.Component {
         <BottomConfirmModal
           isOpen={selection.length > 0}
           questionText={modalStrings.delete_these_invoices}
-          onCancel={() => this.onDeleteCancel()}
-          onConfirm={() => this.onDeleteConfirm()}
+          onCancel={this.onDelete(false)}
+          onConfirm={this.onDelete(true)}
           confirmText={modalStrings.delete}
         />
         <SelectModal
@@ -241,17 +231,15 @@ export class SupplierRequisitionsPage extends React.Component {
           placeholderText={modalStrings.start_typing_to_select_supplier}
           queryString="name BEGINSWITH[c] $0"
           sortByString="name"
-          onSelect={name =>
-            this.setState({ isCreatingRequisition: false }, () => this.onNewRequisition(name))
-          }
-          onClose={() => this.setState({ isCreatingRequisition: false })}
+          onSelect={name => this.onNewRequisition(name)}
+          onClose={this.onNewRequisition}
           title={modalStrings.search_for_the_supplier}
         />
         {byProgramModalOpen && (
           <ByProgramModal
             isOpen={byProgramModalOpen}
             onConfirm={this.onNewRequisition}
-            onCancel={this.onCancelByProgram}
+            onCancel={this.onNewRequisition}
             database={database}
             type="requisition"
             settings={settings}
